[SERVICE-539] Splitting up to prevent 'fin' getting dragged into unit tests
