+++ src/provider/model/DesktopEntity.ts
@@ -2,10 +2,11 @@ import {Point} from 'hadouken-js-adapter/out/types/src/api/system/point';
 
 import {Scope} from '../../../gen/provider/config/layouts-config';
 import {Rectangle} from '../snapanddock/utils/RectUtils';
+import {WindowIdentity} from '../utils/identity';
 
 import {DesktopSnapGroup} from './DesktopSnapGroup';
 import {DesktopTabGroup} from './DesktopTabGroup';
-import {EntityState, WindowIdentity} from './DesktopWindow';
+import {EntityState} from './DesktopWindow';
 
 /**
  * Interface for anything that can be snapped - namely windows and tab sets. Represents any entity that should be
