+++ src/provider/model/DesktopWindow.ts
@@ -1,5 +1,5 @@
 import deepEqual from 'fast-deep-equal';
-import {Identity, Window} from 'hadouken-js-adapter';
+import {Window} from 'hadouken-js-adapter';
 import {WindowInfo} from 'hadouken-js-adapter/out/types/src/api/window/window';
 import {Aggregators, Signal} from 'openfin-service-signal';
 
@@ -14,6 +14,7 @@ import {isWin10} from '../snapanddock/utils/platform';
 import {Point} from '../snapanddock/utils/PointUtils';
 import {Rectangle} from '../snapanddock/utils/RectUtils';
 import {forEachProperty} from '../utils/iteration';
+import {WindowIdentity} from '../utils/identity';
 
 import {DesktopEntity} from './DesktopEntity';
 import {DesktopModel} from './DesktopModel';
@@ -47,11 +48,6 @@ export interface ResizeConstraint {
     resizableMax: boolean;
     minSize: number;
     maxSize: number;
-}
-
-export interface WindowIdentity extends Identity {
-    uuid: string;
-    name: string;
 }
 
 /**
