+++ src/provider/utils/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_TABSTRIP_HEIGHT: number = 60;
