+++ src/provider/workspaces/placeholder.ts
@@ -5,7 +5,7 @@ import {_Window} from 'hadouken-js-adapter/out/types/src/api/window/window';
 import {WorkspaceApp, WorkspaceWindow} from '../../client/workspaces';
 import {model, tabService} from '../main';
 import {DesktopSnapGroup} from '../model/DesktopSnapGroup';
-import {WindowIdentity} from '../model/DesktopWindow';
+import {WindowIdentity, getId} from '../utils/identity';
 import {isWin10} from '../snapanddock/utils/platform';
 
 const DEFAULT_PLACEHOLDER_URL = (() => {
@@ -284,15 +284,6 @@ export function adjustSizeOfFormerlyTabbedWindows(layoutWindow: WorkspaceWindow,
             }
         }
     }
-}
-
-/**
- * Converts an identity into a string ID. Mirrors {@link DesktopModel.getId}.
- *
- * @param identity Any entity identity
- */
-export function getId(identity: WindowIdentity): string {
-    return `${identity.uuid}/${identity.name}`;
 }
 
 // TODO: Make placeholder windows close-able
